Verify migration ledger during deploys
